fix decompress function to work with tar files

+++ anvio/kofam.py
@@ -68,9 +68,13 @@ class KofamSetup(object):
 
 
     def decompress_files(self):
-        print("Not implemented yet")
         for file_name in self.files:
-             full_path = os.path.join(self.pfam_data_dir, file_name)
+            full_path = os.path.join(self.kofam_data_dir, file_name)
+
+            if full_path.endswith("tar.gz"): # extract tar file instead of doing gzip
+                utils.tar_extract_file(full_path, output_file_path = self.kofam_data_dir, keep_original=False)
+            else:
+                utils.gzip_decompress_file(full_path, keep_original=False)
 
              utils.gzip_decompress_file(full_path)
              os.remove(full_path)
